app/requirements.txt
databricks-sdk
fastapi
uvicorn
pydantic
mlflow
tabulate
pyspark